% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.cctm}
\alias{.cctm}
\title{Create a 'conditional' ctm model with random effects as offsets}
\usage{
.cctm(mod, coef, negative = FALSE)
}
\arguments{
\item{mod}{A \code{ctm} model.}

\item{coef}{Coefficient vector for the dummy ctm model.}

\item{negative}{The sign of the random effect term in the corresponding
\code{tramME} model.}
}
\description{
Create a 'conditional' ctm model with random effects as offsets
}
