% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_coef}
\alias{.check_coef}
\title{Check the validity of a vector of coefficients}
\usage{
.check_coef(x, mst)
}
\arguments{
\item{x}{Vector of coefficients}

\item{mst}{List describing the model structure}
}
\description{
Check the validity of a vector of coefficients
}
