% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramTMB.R
\name{.check_par}
\alias{.check_par}
\title{Helper function to check parameter constarints}
\usage{
.check_par(obj, par, eps = 1e-07, ...)
}
\arguments{
\item{obj}{A \code{tramTMB} object}

\item{par}{A parameter vector}

\item{eps}{Tolearnce level}

\item{...}{optional arguments}
}
\description{
Helper function to check parameter constarints
}
