% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_varcov}
\alias{.check_varcov}
\title{Check the validity of a vector of coefficients}
\usage{
.check_varcov(x, vc)
}
\arguments{
\item{x}{List of covariance matrices}

\item{vc}{varcov from the tramME object}
}
\description{
Check the validity of a vector of coefficients
}
