% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramTMB.R
\name{.combine_formulas}
\alias{.combine_formulas}
\title{Combine a set of formulas into one (similar to nlme::asOneFormula)}
\usage{
.combine_formulas(formula, ..., omit = ".")
}
\arguments{
\item{formula}{the first formula, if it contains a response that will be the
response of the resulting formula}

\item{...}{objects from which a formula can be extracted}

\item{omit}{parameter vector with variable names to be omitted}
}
\description{
Combine a set of formulas into one (similar to nlme::asOneFormula)
}
