% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramTMB.R
\name{.constr_adj}
\alias{.constr_adj}
\title{Helper function to adjust the constraints consistently with the parameter
restrictions}
\usage{
.constr_adj(par, constr, map)
}
\arguments{
\item{par}{list containing the vector of parameters}

\item{constr}{list containing the the constraints}

\item{map}{List defining how to optionally collect and fix parameters - see details.}
}
\value{
A list with adjusted constraints.
}
\description{
Helper function to adjust the constraints consistently with the parameter
restrictions
}
