% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramTMB.R
\name{.ctm2formula}
\alias{.ctm2formula}
\title{Create a dummy fromula from a ctm object}
\usage{
.ctm2formula(ctm)
}
\arguments{
\item{ctm}{A ctm model from wich the fromula is created}
}
\description{
Create a dummy fromula from a ctm object
}
