% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.dummy_ctm}
\alias{.dummy_ctm}
\title{Dummy ctm model with random effects as offsets}
\usage{
.dummy_ctm(mst, coef)
}
\arguments{
\item{mst}{List describing the model structure}

\item{coef}{Coefficient vector for the dummy ctm model}
}
\description{
Dummy ctm model with random effects as offsets
}
