% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.gen_param}
\alias{.gen_param}
\title{Generator for param closure}
\usage{
.gen_param(par, fe, re, varnames)
}
\arguments{
\item{par}{Named list of initial parameters (beta and theta).}

\item{fe}{Necessary information about fixed effects.}

\item{re}{Necessary infromation about random effects.}

\item{varnames}{Names of the variables in the model.}
}
\description{
Generator for param closure
}
