% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramTMB.R
\name{.get_par}
\alias{.get_par}
\title{Helper function to extract formatted parameters}
\usage{
.get_par(obj, par = obj$env$par_checked, fixed = TRUE)
}
\arguments{
\item{obj}{A \code{tramTMB} object}

\item{par}{A parameter vector to be formatted}

\item{fixed}{Logica; print fixed parameters, too}
}
\description{
Helper function to extract formatted parameters
}
