% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.idx}
\alias{.idx}
\title{Get parameter indices of various structures}
\usage{
.idx(
  obj,
  fixed = NULL,
  pargroup = "all",
  which = NULL,
  pmatch = FALSE,
  altpar = NULL
)
}
\arguments{
\item{obj}{A tramME object}

\item{fixed}{Logical; should the indices of fixed parameters also be returned?}

\item{pargroup}{Parameter group}

\item{which}{Parameter names or indices within groups}

\item{pmatch}{Is partial matching allowed for \code{which}}

\item{altpar}{Alternative parameterizations (currently only "lm" possible)}
}
\description{
If fixed is logical, it inidcates that the indices refer to the extended
parameter vector
pargroup = c("fixef", "ranef", "shift", "all")
}
