% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramTMB.R
\name{.mod_negative}
\alias{.mod_negative}
\title{Helper function to figure out if negative = TRUE in a given model}
\usage{
.mod_negative(ctm, tram = NULL)
}
\arguments{
\item{ctm}{A \code{ctm} model}

\item{tram}{tram model name: Lm, BoxCox, Colr, Polr, Coxph, Survreg, Lehmann,
Aareg, or the suffixed versions of these (e.g. ColrME). Ignored when a \code{ctm} model
is also supplied.}
}
\description{
Helper function to figure out if negative = TRUE in a given model
}
