% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.model_name}
\alias{.model_name}
\title{Create name for a specific tramME model}
\usage{
.model_name(mst)
}
\arguments{
\item{mst}{model structure list}
}
\description{
Create name for a specific tramME model
}
