% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.nm2mat}
\alias{.nm2mat}
\title{Add names to the vc matrix}
\usage{
.nm2mat(m, rnms)
}
\arguments{
\item{m}{List of covariance matrices without names}

\item{rnms}{Named list of random effects names}
}
\value{
The same list of covariance matrices with proper names
}
\description{
Add names to the vc matrix
}
