% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.nm2vec}
\alias{.nm2vec}
\title{Add names to the vector of theta parameters}
\usage{
.nm2vec(v, rnms)
}
\arguments{
\item{v}{theta vector without names}

\item{rnms}{Named list of random effects names}
}
\value{
The same vector with proper names
}
\description{
Add names to the vector of theta parameters
}
