% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramTMB.R
\name{.nobars}
\alias{.nobars}
\title{Remove random effects terms from full formulas or calls}
\usage{
.nobars(term)
}
\arguments{
\item{term}{Call or formula}
}
\description{
Remove random effects terms from full formulas or calls
}
