% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.parallel_default}
\alias{.parallel_default}
\title{Boilerplate parallel-handling function, modified from \code{glmmTMB}}
\usage{
.parallel_default(parallel = c("no", "multicore", "snow"), ncpus = 1L)
}
\arguments{
\item{parallel}{Parallel backend}

\item{ncpus}{Number of cores/cpus}
}
\description{
Boilerplate parallel-handling function, modified from \code{glmmTMB}
}
