% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.paridx}
\alias{.paridx}
\title{Match parameters with parameter groups and indeces}
\usage{
.paridx(mst, which = NULL, pargroup = "all", pmatch = FALSE, altpar = NULL)
}
\arguments{
\item{mst}{model structure as in an obj$model}

\item{which}{parameter name or index within group}

\item{pargroup}{parameter group}

\item{pmatch}{partrial matching allowed}

\item{altpar}{optional alternative parametrization such as LMM (\code{altpar = "lm"})}
}
\description{
Match parameters with parameter groups and indeces
}
