% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.re_data}
\alias{.re_data}
\title{Create RE data}
\usage{
.re_data(rhsterm, data, negative = FALSE)
}
\arguments{
\item{rhsterm}{Right-hand-side term of the formula}

\item{data}{data}

\item{negative}{should the random effects terms be multiplied with -1?}
}
\description{
Create RE data
}
