% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.re_format}
\alias{.re_format}
\title{Format random effects}
\usage{
.re_format(x, rts, rnms, rbs, rlev)
}
\arguments{
\item{x}{Vector of random effects}

\item{rts}{Vector of length of random effects terms for each grouping factor}

\item{rnms}{Named list of random effects names}

\item{rbs}{Vector of random effects matrix dimensions}

\item{rlev}{List of the levels of each grouping factor}
}
\description{
Format random effects
}
