% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.re_size}
\alias{.re_size}
\title{Calculates the size of the random effect vector implied by the model and the data}
\usage{
.re_size(mst, data)
}
\arguments{
\item{mst}{list describing the model structure}

\item{data}{Dataset containing the required grouping factors}
}
\description{
Calculates the size of the random effect vector implied by the model and the data
}
