% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.renames}
\alias{.renames}
\title{Return names for random effects parameters}
\usage{
.renames(rst)
}
\arguments{
\item{rst}{the random effect structure as saved in obj$model}
}
\description{
Return names for random effects parameters
}
