% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.set_cf}
\alias{.set_cf}
\title{Set the coefficient vector}
\usage{
.set_cf(obj, val)
}
\arguments{
\item{obj}{The tramME object}

\item{val}{The vector of new values}
}
\value{
A new list of parameters with updated beta part
}
\description{
Set the coefficient vector
}
