% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.set_vc}
\alias{.set_vc}
\title{Set the parameters of the random effect covariance matrix}
\usage{
.set_vc(obj, val, as.theta = FALSE)
}
\arguments{
\item{obj}{The tramME object}

\item{val}{The vector of new values}

\item{as.theta}{The input is given according to the reparameterization used by tramTMB}
}
\value{
A new list of parameters with updated beta part
}
\description{
Set the parameters of the random effect covariance matrix
}
