% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{.sim_re}
\alias{.sim_re}
\title{Simulates random effects vector from a tramME object}
\usage{
.sim_re(vc, n)
}
\arguments{
\item{vc}{list of RE variance-covariances}

\item{n}{list of number of values to be simulated for each grouping factor}
}
\description{
Simulates random effects vector from a tramME object
}
