% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.th2vc}
\alias{.th2vc}
\title{Convert from theta vector to vc matrix}
\usage{
.th2vc(th, rbs)
}
\arguments{
\item{th}{Vector of theta parameters (reparametrization of the covariance matrices)}

\item{rbs}{List of random effects matrix dimensions}
}
\description{
Convert from theta vector to vc matrix
}
