% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{.tramME}
\alias{.tramME}
\title{General ME tram model}
\usage{
.tramME(formula, data, na.action, silent, nofit, optim_control)
}
\arguments{
\item{formula}{an object of class \code{"formula"}: a symbolic description 
          of the model structure to be
          fitted.  The details of model specification are given under
          Details and in the package vignette.
  }

\item{data}{an optional data frame, list or environment (or object
          coercible by \code{as.data.frame} to a data frame) containing the
          variables in the model.  If not found in \code{data}, the
          variables are taken from \code{environment(formula)}.
  }

\item{na.action}{a function which indicates what should happen when the data
          contain \code{NA}s.  The default is set to \code{na.omit}.
  }

\item{silent}{Logical, if TRUE, prints all tracing information.}

\item{nofit}{Logical, if TRUE, creates the model objects, but does not run the optimization.}

\item{optim_control}{List of optional arguments for the optimizer.}
}
\description{
General ME tram model
}
