% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramTMB.R
\name{.tramME2ctm}
\alias{.tramME2ctm}
\title{Create a corresponding ctm model for a tramME model}
\usage{
.tramME2ctm(formula, data, mname, ...)
}
\arguments{
\item{formula}{Model formula.}

\item{data}{A \code{data.frame} containing the model variables.}

\item{mname}{tram(ME) model name.}

\item{...}{Optional arguments passed to \code{\link[tram]{tram}}}
}
\description{
Takes a tramME formula and generates the FE ctm model (model_only = TRUE)
Also decides if negative == TRUE
}
