% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.vc2th}
\alias{.vc2th}
\title{Convert from vc matrix to theta vector}
\usage{
.vc2th(vc, rbs)
}
\arguments{
\item{vc}{Covariance matrix of random effects}

\item{rbs}{List of random effects matrix dimensions}
}
\description{
Convert from vc matrix to theta vector
}
