% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramTMB.R
\name{duplicate}
\alias{duplicate}
\title{Generic for copying objects that are (partly) modified in place}
\usage{
duplicate(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Optional parameters.}
}
\description{
Generic for copying objects that are (partly) modified in place
}
