% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{duplicate.tramME}
\alias{duplicate.tramME}
\title{Duplicate a tramME object}
\usage{
\method{duplicate}{tramME}(object, ...)
}
\arguments{
\item{object}{A \code{tramME} object.}

\item{...}{Optional arguments (currently ignored).}
}
\description{
In general, this is not necessary for the usual usage of tramME.
It is only written to avoid errors stemming from the fact that
some parts of the tramME object are modified in place.
}
