% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramTMB.R
\name{duplicate.tramTMB}
\alias{duplicate.tramTMB}
\title{Create a duplicate of the tramTMB object}
\usage{
\method{duplicate}{tramTMB}(object, ...)
}
\arguments{
\item{object}{A \code{tramTMB} object.}

\item{...}{Optional parameters (not used).}
}
\description{
Create a duplicate of the tramTMB object
}
