% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{edf_smooth}
\alias{edf_smooth}
\title{Calculate edf for smooth terms}
\usage{
edf_smooth(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Optional parameters.}
}
\description{
Calculate edf for smooth terms
}
