% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramTMB.R
\name{fe_terms}
\alias{fe_terms}
\title{Create fixed effects data and initial parameters}
\usage{
fe_terms(mod)
}
\arguments{
\item{mod}{a mlt model}
}
\description{
Create fixed effects data and initial parameters
}
