% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitmod.tramME}
\alias{fitmod.tramME}
\title{Call the optimizer on a tramME object}
\usage{
\method{fitmod}{tramME}(object, initpar = NULL, control = optim_control(), ...)
}
\arguments{
\item{object}{A \code{tramME} object.}

\item{initpar}{named list of initial parameter values, if \code{NULL}, it is ignored}

\item{control}{list with controls for optimization}

\item{...}{additional arguments to \code{\link[tram]{tram}}.}
}
\description{
Call the optimizer on a tramME object
}
