% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpterms.R
\name{lpterms}
\alias{lpterms}
\title{Generic method for extracting terms of the linear predictor}
\usage{
lpterms(object, ...)
}
\arguments{
\item{object}{A model object}

\item{...}{Optional parameters}
}
\value{
The value of the baseline transfromation function at certain points.
}
\description{
Generic method for extracting terms of the linear predictor
}
