% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{offset<-}
\alias{offset<-}
\title{Generic method for \code{"offset<-"}}
\usage{
offset(object) <- value
}
\arguments{
\item{object}{A model object.}

\item{value}{The new vector of the offsets.}
}
\value{
An object with the same class as \code{object}, with updated
  offset vector.
}
\description{
Generic method for \code{"offset<-"}
}
