% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{offset.default}
\alias{offset.default}
\title{Default method for \code{"offset"}}
\usage{
\method{offset}{default}(object)
}
\arguments{
\item{object}{An offset to be included in a model frame}
}
\description{
Overloads the original \code{\link[stats]{offset}} function.
}
