% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{offset.tramME}
\alias{offset.tramME}
\title{Get the offset vector of a tramME object.}
\usage{
\method{offset}{tramME}(object)
}
\arguments{
\item{object}{A \code{tramME} object.}
}
\description{
Get the offset vector of a tramME object.
}
