% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramTMB.R
\name{optim_tramTMB}
\alias{optim_tramTMB}
\title{Optimize the tramTMB object}
\usage{
optim_tramTMB(
  obj,
  par = NULL,
  method = "nlminb",
  control = list(),
  trace = FALSE,
  ntry = 5,
  scale = TRUE,
  ...
)
}
\arguments{
\item{obj}{a tramTMB object}

\item{par}{optional vector of initial parameter values}

\item{method}{the method used by \code{alabama::auglag}}

\item{control}{a list of control parameters}

\item{trace}{logical, whether the trace should be printed during the optimization}

\item{ntry}{number of restarts with perturbed initial values when not converged}

\item{scale}{Logical, if \code{TRUE}, the fixed effects design matrices are scaled
to improve convergence}

\item{...}{optional arguments, currently not in use}
}
\description{
Currently only with \code{alabama::auglag} with either \code{nlminb} or \code{optim}
in the case of constrained optimization and \code{nlminb} if there are no constraints.
}
