% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confband.R
\name{plot_ci}
\alias{plot_ci}
\alias{plot.confband.tramME}
\title{Plot confidence bands from \code{tramME} models}
\usage{
\method{plot}{confband.tramME}(
  x,
  col,
  lty,
  fill,
  add = FALSE,
  single_plot = FALSE,
  trafo_x = identity,
  trafo_y = identity,
  align_xlim = FALSE,
  align_ylim = FALSE,
  ...
)
}
\arguments{
\item{x}{The object containing the confidence intervals.}

\item{col}{Color of the point estimates.}

\item{lty}{Line type of the point estimates.}

\item{fill}{Fill color for the intervals.}

\item{add}{If \code{TRUE}, no new plot is created, the interval is added to
the current plot.}

\item{single_plot}{If \code{TRUE}, a single new plot is created, and all
intervals are plotted on it.}

\item{trafo_x}{Transform x-axis before plotting.}

\item{trafo_y}{Transform y-axis before plotting.}

\item{align_xlim}{If \code{TRUE}, align the x-axis limits across all
subplots.}

\item{align_ylim}{If \code{TRUE}, align the y-axis limits across all
subplots.}

\item{...}{Optional arguments passed to \code{\link[graphics]{plot.default}}
and \code{\link[graphics]{plot.xy}}.}
}
\description{
Plotting method for \code{confband.tramME} objects.
}
