% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramTMB.R
\name{predict.tramTMB}
\alias{predict.tramTMB}
\title{Post-estimation calculations in a tramTMB model}
\usage{
\method{predict}{tramTMB}(
  object,
  newdata,
  parameters = .get_par(object, full = TRUE),
  scale = c("lp", "trafo"),
  cov = FALSE,
  as.lm = FALSE,
  ...
)
}
\arguments{
\item{object}{A \code{tramTMB} object}

\item{newdata}{A named list with elements Y, X and Z (not all necessary)}

\item{parameters}{A named list of parameter values}

\item{scale}{The scale on which the post-estimation calculations are done}

\item{cov}{Logical; If \code{TRUE}, calculate the full covariance matrix
of the calculated values}

\item{as.lm}{Logical; reparameterize as a LMM}

\item{...}{Optional arguments (ignored).}
}
\description{
Post-estimation calculations in a tramTMB model
}
