% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.anova.tramME}
\alias{print.anova.tramME}
\title{Printing \code{anova.tramME} table}
\usage{
\method{print}{anova.tramME}(x, digits = max(getOption("digits") - 2L,
  3L), signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{x}{A \code{anova.tramME} object.}

\item{digits}{minimum number of significant digits to be used for
    most numbers.}

\item{signif.stars}{logical; if \code{TRUE}, P-values are additionally
    encoded visually as \sQuote{significance stars} in order to help scanning
    of long coefficient tables.  It defaults to the
    \code{show.signif.stars} slot of \code{\link{options}}.}

\item{...}{Optional arguments passed to \code{\link[stats]{printCoefmat}}}
}
\value{
Invisibly retrurns the \code{anova.tramME} object.
}
\description{
Printing \code{anova.tramME} table
}
