% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{print.simulate.tramME}
\alias{print.simulate.tramME}
\title{Print method for \code{simulate.tramME} objects}
\usage{
\method{print}{simulate.tramME}(x, suppress_seed = TRUE, ...)
}
\arguments{
\item{x}{A \code{simulate.tramME} object}

\item{suppress_seed}{Logical, suppress seed if true.}

\item{...}{Additional parameters passed to various print methods.}
}
\value{
The input simulate.tramME object, invisibly.
}
\description{
Automatically hides the seed attribute of the object.
}
