% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.tramME}
\alias{print.summary.tramME}
\title{Print method for tramME model summary}
\usage{
\method{print}{summary.tramME}(x, digits = max(getOption("digits") - 2L,
  3L), signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{x}{a \code{summary.tramME} object}

\item{digits}{minimum number of significant digits to be used for
    most numbers.}

\item{signif.stars}{logical; if \code{TRUE}, P-values are additionally
    encoded visually as \sQuote{significance stars} in order to help scanning
    of long coefficient tables.  It defaults to the
    \code{show.signif.stars} slot of \code{\link{options}}.}

\item{...}{Optional arguments passed to \code{\link[stats]{printCoefmat}}}
}
\value{
The input summary.tramME object, invisibly.
}
\description{
Print method for tramME model summary
}
