% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.tramME}
\alias{print.tramME}
\title{Print tramME model}
\usage{
\method{print}{tramME}(x, digits = max(getOption("digits") - 2L, 3L), ...)
}
\arguments{
\item{x}{A fitted or unfitted tramME model}

\item{digits}{Number of significant digits}

\item{...}{Optional arguments (for consistency with the generic)}
}
\value{
The original tramME obejct invisibly
}
\description{
Print tramME model
}
