% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.VarCorr.tramME}
\alias{print.VarCorr.tramME}
\title{Print method for the variance-correlation parameters of a tramME object}
\usage{
\method{print}{VarCorr.tramME}(x, sd = TRUE,
  digits = max(getOption("digits") - 2L, 3L), ...)
}
\arguments{
\item{x}{A VarCorr.tramME object}

\item{sd}{Logical. Print standard deviations instead of variances.}

\item{digits}{Number of digits}

\item{...}{optional arguments}
}
\description{
Print method for the variance-correlation parameters of a tramME object
}
