% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{refit.tramME}
\alias{refit.tramME}
\title{Refit the model with a new response vector}
\usage{
\method{refit}{tramME}(object, newresp, ...)
}
\arguments{
\item{object}{A tramME object.}

\item{newresp}{A vector of new response values.}

\item{...}{optional arguments for compatibility}
}
\description{
Useful for parametric bootstrap.
}
