% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trafo.R
\name{trafo}
\alias{trafo}
\title{Generic method for extracting baseline transformations}
\usage{
trafo(object, ...)
}
\arguments{
\item{object}{A model object}

\item{...}{Optional parameters}
}
\value{
The value of the baseline transfromation function at certain points.
}
\description{
Generic method for extracting baseline transformations
}
