% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramTMB.R
\name{tramTMB_inputs}
\alias{tramTMB_inputs}
\title{Create inputs of the tramTMB model}
\usage{
tramTMB_inputs(model, ft, rt, data, param = NULL)
}
\arguments{
\item{model}{a list describing the structure of the model as returned by \code{tramME_model}}

\item{ft}{fixed effects terms as returned by the function \code{fe_terms}}

\item{rt}{random effects terms as returned by the function \code{re_terms}}

\item{data}{model frame containing offsets and weights}

\item{param}{optional named list of initial parameter values}
}
\value{
A list with data matrices and initial parameter values
}
\description{
The function generates random values for \code{NA} values in the list of initial parameter
values.
}
