% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{varcov<-}
\alias{varcov<-}
\title{Generic method for \code{"varcov<-"}}
\usage{
varcov(object, ...) <- value
}
\arguments{
\item{object}{A model object.}

\item{...}{Optional inputs.}

\item{value}{The new value of the covariance matrix.}
}
\value{
An object with the same class as \code{object}, with updated
  variance-covariance matrix of random effects.
}
\description{
Generic method for \code{"varcov<-"}
}
