% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{varcov}
\alias{varcov}
\title{Generic method for \code{varcov}}
\usage{
varcov(object, ...)
}
\arguments{
\item{object}{A model object}

\item{...}{Optional parameters}
}
\value{
A variance-covariance matrix.
}
\description{
Generic method for \code{varcov}
}
