% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramTMB.R
\name{vcov.tramTMB}
\alias{vcov.tramTMB}
\title{Variance-covariance matrix of the parameters}
\usage{
\method{vcov}{tramTMB}(
  object,
  par = object$env$par_checked,
  method = c("optimHess", "numDeriv", "analytical"),
  control = list(),
  ...
)
}
\arguments{
\item{object}{A \code{tramTMB} object.}

\item{par}{An optional vector of parameter values.}

\item{method}{Method for calculating the covariance matrix.}

\item{control}{Optional named list of controls to be passed to the specific methods.}

\item{...}{Optional arguments (ignored)}
}
\description{
Variance-covariance matrix of the parameters
}
