% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{weights<-.tramME}
\alias{weights<-.tramME}
\title{Set the values of the observation weights of a tramME model.}
\usage{
\method{weights}{tramME}(object) <- value
}
\arguments{
\item{object}{A \code{tramME} object defined with \code{do_update = TRUE}.}

\item{value}{A vector of new weight values.}
}
\value{
A  \code{tramME} object with the new weight values.
}
\description{
This method updates the internal \code{tramTMB} object, the \code{model.frame}
of the \code{tramME} object and the function call to propagate the change.
}
\note{
It works only when the \code{tramME} model is defined with \code{do_update = TRUE}.
}
