% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{weights<-}
\alias{weights<-}
\title{Generic method for \code{"weights<-"}}
\usage{
weights(object) <- value
}
\arguments{
\item{object}{A model object.}

\item{value}{The new vector of the weights.}
}
\value{
An object with the same class as \code{object}, with updated
  weight vector.
}
\description{
Generic method for \code{"weights<-"}
}
