% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{weights.tramME}
\alias{weights.tramME}
\title{Get the observation weight vector of a tramME object.}
\usage{
\method{weights}{tramME}(object, ...)
}
\arguments{
\item{object}{A \code{tramME} object.}

\item{...}{Optional arguments (ignored).}
}
\description{
Get the observation weight vector of a tramME object.
}
