% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ColrME}
\alias{ColrME}
\title{ME version of tram::Colr}
\usage{
ColrME(formula, data, subset, weights, offset, na.action = na.omit,
  silent = TRUE, nofit = FALSE, optim_control = list(outer = list(),
  optim = list()), ...)
}
\arguments{
\item{formula}{an object of class \code{"formula"}: a symbolic description 
          of the model structure to be
          fitted.  The details of model specification are given under
          \code{\link{tram}} and in the package vignette.
  }

\item{data}{an optional data frame, list or environment (or object
          coercible by \code{as.data.frame} to a data frame) containing the
          variables in the model.  If not found in \code{data}, the
          variables are taken from \code{environment(formula)}.
  }

\item{subset}{an optional vector specifying a subset of observations to be
          used in the fitting process.
  }

\item{weights}{an optional vector of weights to be used in the fitting
          process.  Should be \code{NULL} or a numeric vector. If present,
          the weighted log-likelihood is maximised.
  }

\item{offset}{this can be used to specify an _a priori_ known component to
          be included in the linear predictor during fitting.  This
          should be \code{NULL} or a numeric vector of length equal to the
          number of cases.
  }

\item{na.action}{a function which indicates what should happen when the data
          contain \code{NA}s.  The default is set by the \code{na.action} setting
          of \code{options}, and is \code{na.fail} if that is unset.
  }

\item{silent}{Logical, if TRUE, prints all tracing information.}

\item{nofit}{Logical, if TRUE, creates the model objects, but does not run the optimization.}

\item{optim_control}{List of optional arguments for the optimizer.}

\item{...}{additional arguments to \code{\link{tram}}.}
}
\description{
ME version of tram::Colr
}
