% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.model_name}
\alias{.model_name}
\title{Generates proper model name for the tramME model}
\usage{
.model_name(obj)
}
\arguments{
\item{obj}{A \code{tramME} object.}
}
\description{
Generates proper model name for the tramME model
}
