% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.re_format}
\alias{.re_format}
\title{Reshape unformatted random effects vectors}
\usage{
.re_format(rst, x)
}
\arguments{
\item{rst}{the random effect structure as saved in obj$model}

\item{x}{Unformatted random effects vector}
}
\description{
Reshape unformatted random effects vectors
}
