% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.re_size}
\alias{.re_size}
\title{Calculates the size of the random effect vector implied by the model and the data}
\usage{
.re_size(bs, data)
}
\arguments{
\item{bs}{Blocksize vector as returned by the function \code{re_terms}.}

\item{data}{Dataset containing the required grouping factors.}
}
\description{
Calculates the size of the random effect vector implied by the model and the data
}
