% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.tramME}
\alias{logLik.tramME}
\title{Get the log-likelihood of the model}
\usage{
\method{logLik}{tramME}(object, ...)
}
\arguments{
\item{object}{A fitted tramME model}

\item{...}{Optional argument (for consistency with generic)}
}
\value{
A numeric value of the log-likelihood at its optimum.
}
\description{
Get the log-likelihood of the model
}
\examples{
data("sleepstudy", package = "lme4")
fit <- LmME(Reaction ~ Days + (Days | Subject), data = sleepstudy)
logLik(fit)
}
