% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{model.frame.tramME}
\alias{model.frame.tramME}
\title{Extract model frame from a tramME model}
\usage{
\method{model.frame}{tramME}(formula, ...)
}
\arguments{
\item{formula}{A \code{tramME} object}

\item{...}{Optional arguments (currently ignored)}
}
\description{
Extract model frame from a tramME model
}
