% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.tramME}
\alias{summary.tramME}
\title{Summary method for tramME model}
\usage{
\method{summary}{tramME}(object, ...)
}
\arguments{
\item{object}{A tramME object}

\item{...}{Optional arguments (for consistency with the generic)}
}
\description{
Summary method for tramME model
}
