% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{variable.names.tramME}
\alias{variable.names.tramME}
\title{Get the baseline transformation function and its confidence
interval}
\usage{
\method{variable.names}{tramME}(object, type = c("all", "response",
  "group"), ...)
}
\arguments{
\item{object}{a tramME object (fitted or unfitted)}

\item{type}{all: all non-eliminated variable names,
response: response variable,
group: grouping factors for random effects.}

\item{...}{optional parameters}
}
\value{
A vector of variable names.
}
\description{
For stratified models, it returns a list of data frames for each
  evaluated on.
Return variable names.
}
\details{
Returns the variable names corresponding the selected group.
The returned names are derived names as tramME uses them. For example, when the
response is a Surv object, \code{variable.names} returns the name of that object, and
the names of the variables used to create it.
}
\examples{
data("sleepstudy", package = "lme4")
mod <- LmME(Reaction ~ Days + (Days | Subject), data = sleepstudy, nofit = TRUE)
variable.names(mod)
variable.names(mod, "response")
}
